iOS-#67 [계산기] Meterial Search View 구조 수정
