+++ Amoogye/Common/MeterialSearchView/MeterialSearchView.xib
@@ -10,9 +10,14 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="MeterialSearchView" customModule="Amoogye" customModuleProvider="target"/>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="MeterialSearchView" customModule="Amoogye" customModuleProvider="target">
+            <connections>
+                <outlet property="searchBar" destination="0B4-0D-wQC" id="eWl-o0-n57"/>
+                <outlet property="tableView" destination="Sfh-uq-pGz" id="vSB-7m-COL"/>
+            </connections>
+        </placeholder>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB" customClass="SearchView" customModule="Amoogye" customModuleProvider="target">
+        <view contentMode="scaleToFill" id="iN0-l3-epB">
             <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
@@ -36,10 +41,6 @@
                 <constraint firstItem="Sfh-uq-pGz" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="naE-px-Qzo"/>
             </constraints>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <connections>
-                <outlet property="searchBar" destination="0B4-0D-wQC" id="nRo-Oe-bKW"/>
-                <outlet property="tableView" destination="Sfh-uq-pGz" id="lPz-aS-8pA"/>
-            </connections>
         </view>
     </objects>
 </document>
